Yet another TeXmath-to-HTML gem. This one uses KaTeX and ExecJS.
This is a TeX-to-HTML+MathML+CSS converter class using the Javascript-based
KaTeX, interpreted by one of the Javascript engines supported by ExecJS.
The intended purpose is to eliminate the need for math-rendering JavaScript
in the HTML browser.

Consider this a lightweight alternative to mathjax-node-cli.
Javascript execution context initialization can be done once and then
reused for formula renderings with the same general configuration.
As a result, the performance is reasonable.

Originally added directly to a kramdown fork. However, relying on config
from potentially untrusted sources to find the necessary Javascript files
might open security vulnerabilities. Therefore we chose to make that
feature upgrade an opt-in requiring dedicated admin action, namely the
explicit installation of this gem.